+++ packages/contracts/contracts/test/BalanceManager.t.sol
@@ -82,12 +82,14 @@ contract BalanceManagerTest is Test {
         token.reserveTokens(recipient, amount);
 
         StealthAddress memory addr = NocturneUtils.defaultStealthAddress();
-        Deposit memory deposit = NocturneUtils.formatDeposit(
+        DepositRequest memory deposit = NocturneUtils.formatDepositRequest(
             recipient,
             address(token),
             amount,
             ERC20_ID,
-            addr
+            addr,
+            0, // nonce and gasPrice irrelevant
+            0
         );
 
         vm.prank(recipient);