+++ packages/contracts/contracts/test/Wallet.t.sol
@@ -91,12 +91,12 @@ contract WalletTest is Test, ParseUtils, ForgeUtils, PoseidonDeployer {
             address(subtreeUpdateVerifier)
         );
 
+        vault.initialize(address(wallet));
+
         hasherT3 = IHasherT3(new PoseidonHasherT3(poseidonT3));
         hasherT6 = IHasherT6(new PoseidonHasherT6(poseidonT6));
 
         treeTest.initialize(hasherT3, hasherT6);
-
-        vault.initialize(address(wallet));
 
         // Instantiate token contracts
         for (uint256 i = 0; i < 3; i++) {
@@ -108,19 +108,21 @@ contract WalletTest is Test, ParseUtils, ForgeUtils, PoseidonDeployer {
 
     function depositFunds(
         address _spender,
-        address _asset,
+        address _token,
         uint256 _value,
         uint256 _id,
         StealthAddress memory _depositAddr
     ) public {
         wallet.depositFunds(
-            Deposit({
-                spender: _spender,
-                encodedAssetAddr: uint256(uint160(_asset)),
-                encodedAssetId: _id,
-                value: _value,
-                depositAddr: _depositAddr
-            })
+            NocturneUtils.formatDepositRequest(
+                _spender,
+                _token,
+                _value,
+                _id,
+                _depositAddr,
+                0, // nonce and gasPrice irrelevant once past deposit manager
+                0
+            )
         );
     }
 
@@ -131,7 +133,7 @@ contract WalletTest is Test, ParseUtils, ForgeUtils, PoseidonDeployer {
     ) internal {
         token.reserveTokens(recipient, amount);
 
-        vm.prank(recipient);
+        vm.startPrank(recipient);
         token.approve(address(vault), amount);
 
         uint256[] memory batch = new uint256[](16);
@@ -154,7 +156,6 @@ contract WalletTest is Test, ParseUtils, ForgeUtils, PoseidonDeployer {
                 uint128(i)
             );
 
-            vm.prank(recipient);
             if (i == depositIterations - 1 && remainder != 0) {
                 depositFunds(
                     recipient,
@@ -186,6 +187,8 @@ contract WalletTest is Test, ParseUtils, ForgeUtils, PoseidonDeployer {
             batch[i] = noteCommitment;
         }
 
+        vm.stopPrank();
+
         uint256[] memory path = treeTest.computeInitialRoot(batch);
         uint256 root = path[path.length - 1];
 
@@ -201,15 +204,24 @@ contract WalletTest is Test, ParseUtils, ForgeUtils, PoseidonDeployer {
         vm.prank(ALICE);
         token.approve(address(vault), PER_NOTE_AMOUNT);
 
-        vm.prank(BOB); // msg.sender made to BOB not ALICE, causing error
+        vm.startPrank(BOB); // msg.sender made to BOB not ALICE, causing error
         vm.expectRevert("Spender must be the sender");
-        depositFunds(
-            ALICE,
-            address(token),
-            PER_NOTE_AMOUNT,
-            ERC20_ID,
-            NocturneUtils.defaultStealthAddress()
+        wallet.depositFunds(
+            DepositRequest({
+                chainId: 0,
+                spender: ALICE,
+                encodedAsset: AssetUtils.encodeAsset(
+                    AssetType.ERC20,
+                    address(token),
+                    ERC20_ID
+                ),
+                value: PER_NOTE_AMOUNT,
+                depositAddr: NocturneUtils.defaultStealthAddress(),
+                nonce: 0,
+                gasCompensation: 0
+            })
         );
+        vm.stopPrank();
     }
 
     function testProcessBundleTransferSingleJoinSplit() public {